feat:update branch donation-flow

+++ src/main/java/com/example/blood_donation_support_system/entity/EmergencyEntity.java
@@ -17,7 +17,7 @@ public class EmergencyEntity {
     @Column(name="component_type")
     private String componentType;
     @Column(name="quantity")
-    private double quantity;
+    private Integer quantity;
     @Column(name="note")
     private String note;
     @Column(name="needed_time")
